+++ src/sdks/ignite.ts
@@ -3,6 +3,7 @@ import {API, Id, assertId} from '../rest/index.ts';
 import {
 	Deployment,
 	DeploymentConfig,
+	DeploymentMetaData,
 	Gateway,
 	GatewayType,
 	RuntimeType,
@@ -46,13 +47,22 @@ export const ignite = sdk(client => {
 			return igniteSDK.containers.create(this.id);
 		},
 
-		createGateway(config: {
-			type: API.Ignite.GatewayType;
-			protocol: API.Ignite.Gateway['protocol'];
-			port: number;
-			name: string;
-			targetPort: number;
-		}) {
+		createGateway(
+			config:
+				| {
+						type: API.Ignite.GatewayType.EXTERNAL;
+						protocol: API.Ignite.Gateway['protocol'];
+						name: string;
+						targetPort: number;
+				  }
+				| {
+						type: API.Ignite.GatewayType.INTERNAL;
+						protocol: API.Ignite.Gateway['protocol'];
+						name: string;
+						targetPort: number;
+						internalDomain: string;
+				  },
+		) {
 			return igniteSDK.gateways.create(this.id, config);
 		},
 
@@ -268,21 +278,36 @@ export const ignite = sdk(client => {
 		 */
 		async create(
 			deployment: Deployment | Deployment['id'],
-			config: {
-				type: GatewayType;
-				protocol: Gateway['protocol'];
-				targetPort: number;
-				name: string;
-			},
+			config:
+				| {
+						type: GatewayType.EXTERNAL;
+						protocol: Gateway['protocol'];
+						targetPort: number;
+						name: string;
+				  }
+				| {
+						type: GatewayType.INTERNAL;
+						protocol: Gateway['protocol'];
+						targetPort: number;
+						name: string;
+						internalDomain: string;
+				  },
 		) {
 			const deploymentId =
 				typeof deployment === 'object' ? deployment.id : deployment;
 
-			const {targetPort, ...rest} = config;
+			const body =
+				config.type === GatewayType.EXTERNAL
+					? {...config, target_port: config.targetPort}
+					: {
+							...config,
+							target_port: config.targetPort,
+							internal_domain: config.internalDomain,
+					  };
 
 			const {gateway} = await client.post(
 				'/v1/ignite/deployments/:deployment_id/gateways',
-				{...rest, target_port: targetPort},
+				body,
 				{deployment_id: deploymentId},
 			);
 
@@ -450,6 +475,19 @@ export const ignite = sdk(client => {
 					undefined,
 					{deployment_id: deployment},
 				);
+			},
+
+			async patchMetadata(
+				deploymentId: Id<'deployment'>,
+				metadata: Partial<DeploymentMetaData>,
+			) {
+				const {deployment} = await client.patch(
+					'/v1/ignite/deployments/:deployment_id/metadata',
+					metadata,
+					{deployment_id: deploymentId},
+				);
+
+				return deployment;
 			},
 
 			/**